Boss card component refactored
