--- src/App.css
@@ -1,10 +1,0 @@
-.bosses-container {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-}
-
-.boss-card {
-  width: 300px;
-  margin-bottom: 20px;
-}
