+++ src/assets/images/low_chance.png
